RUST-1881 Check integer conversions (#1045)

+++ src/operation.rs
@@ -71,9 +71,9 @@ const SERVER_4_2_0_WIRE_VERSION: i32 = 8;
 const SERVER_4_4_0_WIRE_VERSION: i32 = 9;
 // The maximum number of bytes that may be included in a write payload when auto-encryption is
 // enabled.
-const MAX_ENCRYPTED_WRITE_SIZE: u64 = 2_097_152;
+const MAX_ENCRYPTED_WRITE_SIZE: usize = 2_097_152;
 // The amount of overhead bytes to account for when building a document sequence.
-const COMMAND_OVERHEAD_SIZE: u64 = 16_000;
+const COMMAND_OVERHEAD_SIZE: usize = 16_000;
 
 /// A trait modeling the behavior of a server side operation.
 ///
